+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Show Avatar Builder for Android` menu to upload an avatar for Android build target with NDMF.
 - Non-destructive avatar conversion with NDMF and `VQT Avatar Converter Settings` component.
 - `VQT Platform Target Settings` component to specify the target platform for `VQT Platform Component Remover` and `VQT Platform GameObject Remover` components.
+- Warning message for using unsupported texture formats.
 
 ### Changed
 - Require NDMF 1.3.0 or later when using NDMF related features.